Fix long

+++ core/src/main/kotlin/com/alecstrong/sql/psi/core/SqlFileBase.kt
@@ -110,7 +110,7 @@ abstract class SqlFileBase(
     block: (SchemaContributor) -> Unit,
   ) {
     if (includeAll) {
-      val orderedContributors = sortedMapOf<Int, LinkedHashSet<SchemaContributor>>()
+      val orderedContributors = sortedMapOf<Long, LinkedHashSet<SchemaContributor>>()
       val topContributors = LinkedHashSet<SchemaContributor>()
       val index = SchemaContributorIndex.getInstance(project)
